Fix/get empty file

+++ src/event/mode/ReadFile.cpp
@@ -33,8 +33,7 @@ void ReadFile::Run(intptr_t offset) {
     char buf[BUF_SIZE];
     int  fd = polled_fd_;
 
-    int read_size = read(fd, buf, BUF_SIZE);
-    // TODO: エラー処理
+    ssize_t read_size = read(fd, buf, BUF_SIZE);
     if (read_size == -1) {
         perror("read");
         throw status::server_error;
